Make dupekit a pure Rust workspace
