+++ lib/marin/src/marin/processing/classification/__init__.py
@@ -11,6 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-# TODO (rav): remove F403 when we stabilize dupekit
-from .dupekit import *  # noqa: F403
